pyproject.toml
[tool.poetry]
name = "job-tracker"
version = "0.1.0"
description = ""
authors = ["Darko Trifunovski <dtrifuno@gmail.com>"]
readme = "README.md"
license = "MIT"

repository = "https://github.com/dtrifuno/job-tracker"
homepage = "https://github.com/dtrifuno/job-tracker"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
graphene = "^2.1.8"
sqlalchemy = "^1.3.16"
graphene-sqlalchemy = "^2.2.2"
flask_sqlalchemy = "^2.4.1"
flask_graphql = "^2.0.1"
flask-graphql-auth = "^1.3.0"
flask-migrate = "^2.5.3"
flask_graphql_auth = "^1.3.0"
flask_cors = "^3.0.8"
validators = "^0.14.3"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
autopep8 = "^1.5.2"

[tool.poetry.scripts]
dev = "scripts:dev"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
